requirements.txt
Django==3.1
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-tablefield==0.0.1b0
djangorestframework==3.11.1
djangotoolbox==1.8.0
gunicorn==20.0.4
heroku==0.1.4
whitenoise==5.2.0
